tg_bot/requirements.txt
python-telegram-bot==13.13
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.40
psycopg2-binary==2.9.3